wall fix?

+++ include/game.h
@@ -6,7 +6,7 @@
 /*   By: zyasuo <zyasuo@student.21-school.ru>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/06/21 16:26:25 by zyasuo            #+#    #+#             */
-/*   Updated: 2022/07/01 23:16:54 by zyasuo           ###   ########.fr       */
+/*   Updated: 2022/07/03 17:50:28 by zyasuo           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -77,5 +77,5 @@ void		clear_player(t_player *player);
 void		draw_ray(t_game *game, t_data *img, int angle);
 void		draw_square(int color, t_data img, float x, float y);
 void		render_rays(t_game *game, t_data *img);
-int			is_wall(float x, float y, char **map);
+int			is_wall(int x, int y, char **map);
 #endif
